Flyttet logik ut ur routes til service-klasse for korrigering IM renamet OMS til KorrigeringIm Flyttet ut URLs objekt til felles objekt (samme for alle routes)
